refactor: split ui prompt api
